+++ index.js
@@ -4,3 +4,4 @@ require('./lib/proxyShim');
 
 exports.connect = require('./lib/db').connect;
 exports.Document = require('./lib/document');
+exports.EmbeddedDocument = require('./lib/embedded-document');